gui format changed to java.swing. basic main menu screen created.
The program has three tabs, settings, distribution, and graph. Settings
is 100% done (I believe). Just need to add listeners to its buttons/text
areas. The large white box taking up half the screen is a scroll pane,
it can be populated with all of the messages that the program outputs to
the user. Graph/distribution tabs were created but contain nothing right
now.